backend/requirements.txt
fastapi==0.121.1
uvicorn==0.34.0
gunicorn==23.0.0
anyio==4.7.0
pydantic==2.12.4
python-dotenv==1.2.1
python_dateutil==2.9.0.post0
python-multipart==0.0.9
SQLAlchemy==2.0.44
sqlmodel==0.0.27
psycopg2-binary==2.9.10
passlib==1.7.4
PyJWT==2.10.1
email-validator==2.2.0
pytest
pytest-asyncio==0.23.6
httpx==0.27.2
